fixed datatypes in main

+++ main.py
@@ -7,8 +7,8 @@ objEMF = EMF()
 objEMF.init_const(1.)
 #objEMF.init_wave(1., 100., (0., 0., 1.), 0.)
 
-m = 1.
-e = 1.
+m = np.double(1.)
+e = np.double(1.)
 
 x0 = np.array([0., 1., 0.])
 p0 = np.array([1., 0., 1.])
